code code code

+++ views/listentity.php
@@ -15,13 +15,28 @@
 
             <?php foreach ($allentities as $entity) {
 
-                echo "<div class=\"col s3 content-container__listbox\">";
+                echo "<div class=\"col s12 m6 l3 content-container__listbox\">";
                 echo "<h5>".$entity->getTitle()."</h5>";
-                echo "<img src=\"".$entity->getThumbnail()."\" alt=\"name\" title=\"THE WRONG X\" class=\"show-super-product\">";
+                if (array_key_exists($entity->getAssetid(),$cartitems)) {
+                    echo "<a href='https://github.com/digital3mpire/SUPER-INFORMATION-HIGH-MARKET/tree/master/'>
+                        <img src=\"".$entity->getThumbnail()."\" alt=\"name\" class=\"selected\" title=\"THE WRONG X\">
+                        </a>";
+                } else {
+                    echo "<a href='https://github.com/digital3mpire/SUPER-INFORMATION-HIGH-MARKET/tree/master/'>
+                        <img src=\"".$entity->getThumbnail()."\" alt=\"name\" title=\"THE WRONG X\">
+                        </a>";
+                }
                 echo "<div class=\"content-container__listbox-subheader\">".$entity->getArtistname()."</div>";
                 echo "<div class=\"content-container__listbox-tools right-align\">";
-                echo "<a href=\"/Supershop/removefromcart/id/".$entity->getAssetid()."\"><i class=\"material-icons\">remove_shopping_cart</i></a>";
-                echo "<a href=\"/Supershop/addtocart/id/".$entity->getAssetid()."\"><i class=\"material-icons\">add_shopping_cart</i></a>";
+
+                if (array_key_exists($entity->getAssetid(),$cartitems)) {
+                    echo "<a href=\"/Supershop/removefromcart/id/".$entity->getAssetid()."\"><i class=\"material-icons\">remove_shopping_cart</i></a>";
+                    echo "<i class=\"material-icons lighten \">add_shopping_cart</i>";
+                } else {
+                    echo "<i class=\"material-icons lighten\">remove_shopping_cart</i>";
+                    echo "<a href=\"/Supershop/addtocart/id/".$entity->getAssetid()."\"><i class=\"material-icons\">add_shopping_cart</i></a>";
+                }
+
                 echo "<i class=\"material-icons show-super-product\">queue</i>";
                 echo "</div>";
                 echo "</div>";
